Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -3,8 +3,8 @@
 
 
 int main() {
-    double a;
-scanf("%ld",&a);
-printf("You entered: %ld",a);
+    float a;
+scanf("%f",&a);
+printf("You entered: %.2f",a);
     return 0;
 }
